parameterize demo notebook and allow float exploration constant
Update the demo notebook to use environment variables for domain configuration and logging, replacing hardcoded values to make the experiments more generic. Additionally, modify the parser to support floating-point values for the MCTS exploration constant to allow for finer search tuning.

+++ unified_planning/parser.py
@@ -39,7 +39,7 @@ parser.add_argument(
     default='mcts',
 )
 parser.add_argument('-do', '--domain', help='domain')
-parser.add_argument('-e', '--exploration_constant', help='the exploration constant for mcts solver', nargs='?', default=10, type=int)
+parser.add_argument('-e', '--exploration_constant', help='the exploration constant for mcts solver', nargs='?', default=10.0, type=float)
 parser.add_argument('-ge', '--garbage_amount', help='how many garbage actions to add to the domain', nargs='?', default=0, type=int)
 parser.add_argument('-oe', '--object_amount', help='how many different objects in the domain', nargs='?', default=1, type=int)
 parser.add_argument('-k', '--k', help='K random actions in the max planner', nargs='?', default=10, type=int)
